Improve quickstart tutorial
Our current Quickstart tutorial was too concise when explaining how to quickstart an agent, and too lengthy when explaining how to do other things.

I have divided the current `agent.md` into two files: `quickstart.md` which only explains how to get a local agent and how to use it.

In a future PR I will add the second part - in a separate file - explaining how to use the registry to upload and download agents.